Refactor event joining logic to use authenticated member ID, enhance reservation validation for loyalty tokens, and update event date type in migration.

+++ database/migrations/2026_03_29_130201_create_loyalty_txns_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('reference_id');
             $table->enum('reference_type', ['Reservation', 'Event']);
             $table->enum('txn_type', ['RESERVATION', 'EVENT']);
-            $table->decimal('points', 8, 2);
+            $table->integer('points');
             $table->string('descriptions')->nullable();
             $table->timestamps();
         });
